vitalguard-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
scikit-learn==1.4.2
numpy==1.26.4
pymongo==4.7.1
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2
twilio==9.0.4

# Uncomment for Firebase:
# firebase-admin==6.5.0
requests==2.31.0
